+++ accounting-service/accounting-api/src/main/java/com/microservices/accounting/api/dto/CardInfo.java
@@ -1,4 +1,4 @@
-package com.microservices.accounting.dto;
+package com.microservices.accounting.api.dto;
 
 public enum CardInfo {
     VALID,
