+++ accounting-service/accounting-api/src/main/java/com/microservices/accounting/api/dto/InvokePaymentDto.java
@@ -1,4 +1,4 @@
-package com.microservices.accounting.dto;
+package com.microservices.accounting.api.dto;
 
 import lombok.Getter;
 import lombok.Setter;
