+++ accounting-service/accounting-api/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+#The name for accounting service that is used by Eureka and Ribbon (in Feign)
+accounting.service.name=accounting-service
