+++ accounting-service/accounting-app/gradle.properties
@@ -1,2 +1,3 @@
 applicationVersion=0.1-SNAPSHOT
 releaseTrainVersion=Greenwich.SR2
+accountingApiVersion=0.1-SNAPSHOT
