+++ accounting-service/accounting-app/src/main/java/com/microservices/accounting/entity/PaymentEntity.java
@@ -1,7 +1,8 @@
 package com.microservices.accounting.entity;
 
-import com.microservices.accounting.dto.InvokePaymentDto;
-import com.microservices.accounting.dto.PaymentStatus;
+import com.microservices.accounting.api.dto.InvokePaymentDto;
+import com.microservices.accounting.api.dto.PaymentDetailsDto;
+import com.microservices.accounting.api.dto.PaymentStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -41,5 +42,10 @@ public class PaymentEntity {
         this.amount = invokePaymentDto.getAmountOfMoney();
         this.username = invokePaymentDto.getUser().getLogin();
         this.paymentStatus = Objects.requireNonNull(status);
+    }
+
+
+    public PaymentDetailsDto toPaymentDetailsDto() {
+        return new PaymentDetailsDto(paymentId, username, amount, paymentStatus);
     }
 }
