+++ accounting-service/accounting-app/src/main/resources/application.properties
@@ -1,6 +1,7 @@
 # The port that will be used for REST access
 server.port=8093
-
+# Application name is used by Eureka Service Discovery
+spring.application.name=accounting-service
 # Database and Hibernate properties
 database.schema.name=account_management
 spring.datasource.url=jdbc:postgresql://localhost:5432/laundry_db
